[Refactor] plan 형식 컴포넌트 분리
